requirements.txt
asgiref==3.8.1
blessed==1.20.0
bpython==0.24
certifi==2024.8.30
charset-normalizer==3.4.0
curtsies==0.4.2
cwcwidth==0.1.9
dj-database-url==2.2.0
Django==4.2.16
django-cors-headers==4.5.0
djangorestframework==3.15.2
exceptiongroup==1.2.2
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
iniconfig==2.0.0
Markdown==3.7
packaging==24.1
pillow==11.0.0
pluggy==1.5.0
psycopg2-binary==2.9.10
Pygments==2.18.0
PyJWT==2.9.0
pytest==8.3.3
pyxdg==0.28
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tomli==2.0.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
whitenoise==6.7.0
